+++ Sources/StreamChatUI/Components.swift
@@ -166,6 +166,10 @@ public struct Components {
     /// The injector used for injecting file attachment views
     public var filesAttachmentInjector: AttachmentViewInjector.Type = FilesAttachmentViewInjector.self
 
+    /// The injector used to combine multiple types of attachment views.
+    /// By default, it is a combination of a file injector and a gallery injector.
+    public var mixedAttachmentInjector: MixedAttachmentViewInjector.Type = MixedAttachmentViewInjector.self
+
     /// The button for taking an action on attachment being uploaded.
     public var attachmentActionButton: AttachmentActionButton.Type = AttachmentActionButton.self