+++ Sources/StreamChatUI/Deprecations.swift
@@ -231,3 +231,44 @@ extension DateFormatter {
         DefaultMessageDateSeparatorFormatter().dateFormatter
     }
 }
+
+// MARK: AttachmentsPreviewVC - Mixed Attachments Support
+
+extension AttachmentsPreviewVC {
+    @available(
+        *,
+        deprecated,
+        message: "this view has been split into 2 views, horizontalScrollView and verticalScrollView. This change was required to support mixed attachments. We highly recommend stopping using this view since with mixed attachments the customization done to this view won't affect both scroll views."
+    )
+    open var scrollView: UIScrollView {
+        let axises = Set(content.map { type(of: $0).preferredAxis })
+        if axises.contains(.horizontal) {
+            return horizontalScrollView
+        }
+        return verticalScrollView
+    }
+
+    @available(*, deprecated, renamed: "verticalScrollViewHeightConstraint")
+    public var scrollViewHeightConstraint: NSLayoutConstraint? {
+        get { verticalScrollViewHeightConstraint }
+        set { verticalScrollViewHeightConstraint = newValue }
+    }
+
+    @available(
+        *,
+        deprecated,
+        message: "this property is not being used anymore by default. There's now two scroll views for each axis."
+    )
+    open var stackViewAxis: NSLayoutConstraint.Axis {
+        content.first.flatMap { type(of: $0).preferredAxis } ?? .horizontal
+    }
+
+    @available(
+        *,
+        deprecated,
+        message: "it has been replaced by attachmentPreviews(for:). The name reflects better the intent and when asking for the attachment views, it is safer to specify which axis or axises we want."
+    )
+    open var attachmentViews: [UIView] {
+        attachmentPreviews(for: [.horizontal, .vertical])
+    }
+}