+++ StreamChatUITestsAppUITests/Pages/StartPage.swift
@@ -1,0 +1,10 @@
+//
+// Copyright © 2022 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+import XCTest
+
+class StartPage {
+    static var startButton: XCUIElement { app.buttons["TestApp.Start"] }
+}