+++ Tests/StreamChatTests/APIClient/Endpoints/MessageEndpoints_Tests.swift
@@ -138,14 +138,10 @@ final class MessageEndpoints_Tests: XCTestCase {
             method: .post,
             queryItems: nil,
             requiresConnectionId: false,
-            body: [
-                "reaction": MessageReactionRequestPayload(
-                    type: reaction,
-                    score: score,
-                    enforceUnique: false,
-                    extraData: extraData
-                )
-            ]
+            body: MessageReactionRequestPayload(
+                enforceUnique: false,
+                reaction: ReactionRequestPayload(type: reaction, score: score, extraData: extraData)
+            )
         )
         
         // Build endpoint.
@@ -215,5 +211,24 @@ final class MessageEndpoints_Tests: XCTestCase {
         // Assert endpoint is built correctly
         XCTAssertEqual(AnyEndpoint(expectedEndpoint), AnyEndpoint(endpoint))
         XCTAssertEqual("messages/\(messageId)/action", endpoint.path.value)
+    }
+    
+    func test_translate_buildsCorrectly() {
+        let messageId: MessageId = .unique
+        let language = TranslationLanguage.allCases.randomElement()!
+        
+        let expectedEndpoint = Endpoint<MessagePayload.Boxed>(
+            path: .translateMessage(messageId),
+            method: .post,
+            queryItems: nil,
+            requiresConnectionId: false,
+            body: ["language": language.languageCode]
+        )
+        
+        let endpoint: Endpoint<MessagePayload.Boxed> = .translate(messageId: messageId, to: language)
+        
+        // Assert endpoint is built correctly
+        XCTAssertEqual(AnyEndpoint(expectedEndpoint), AnyEndpoint(endpoint))
+        XCTAssertEqual("messages/\(messageId)/translate", endpoint.path.value)
     }
 }