+++ Tests/StreamChatUITests/SnapshotTests/Composer/ComposerVC_Tests.swift
@@ -158,6 +158,46 @@ class ComposerVC_Tests: XCTestCase {
         
         AssertSnapshot(composerVC)
     }
+
+    func test_attachmentsPreview_withMultipleAttachmentTypes() {
+        let composerVC = ComposerVC()
+        composerVC.appearance = Appearance.default
+
+        composerVC.content = .initial()
+        composerVC.content.attachments = [.mockFile, .mockFile, .mockImage, .mockVideo]
+
+        AssertSnapshot(composerVC)
+    }
+    
+    func test_commandWithNonEmptyArgs_hasSendButtonDisabled() {
+        let composerVC = ComposerVC()
+        composerVC.appearance = Appearance.default
+        
+        composerVC.content = .initial()
+        composerVC.content.addCommand(.init(
+            name: "ARGNEEDED",
+            description: "some command",
+            set: "special",
+            args: "[text]"
+        ))
+        
+        AssertSnapshot(composerVC)
+    }
+    
+    func test_commandWithEmptyArgs_hasSendButtonEnabled() {
+        let composerVC = ComposerVC()
+        composerVC.appearance = Appearance.default
+        
+        composerVC.content = .initial()
+        composerVC.content.addCommand(.init(
+            name: "NOARGNEEDED",
+            description: "some command",
+            set: "special",
+            args: ""
+        ))
+        
+        AssertSnapshot(composerVC)
+    }
     
     func test_whenSuggestionsLookupIsLocal_onlyChannelMembersAreShown() {
         final class ComposerContainerVC: UIViewController {